requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
sqlalchemy==2.0.23
pydantic==2.5.2
python-multipart==0.0.6
python-dotenv==1.0.0
cloudinary==1.36.0
Pillow==10.1.0
qrcode==7.4.2
requests==2.31.0
pg8000==1.30.3